Update community.js
Currently there isn't a simple what to sign up for the Pulsar Slack.  The link to the heroku app seems to have been lost in the move to the new site.  
I used this page as the model for the text: https://pulsar.apache.org/contact/

approx line 249- Added text to explain the new button
approx line 253- Add button for the heroku app and the text explaining the button
approx line 233- correcting the url
approx line 200- removed unnecessary prepositional phrase "which includes Pulsar committers".. yes a minor change but since I was adding in the button...