Easier to copy&paste command for Windows
I know its something small but it really should be able for windows people to simply copy and paste this shell command without needing to do any extra effort like removing the backslashes and making it one line because powershell doesn't support this. If you disagree feel free to decline this.