Change font-size in docs from 18px to 16px (#814)
The purpose of this change is to increase the amount of content that fits on the screen.

The lack of horizontal space makes it hard to read table content on documentation pages.

Discussion: apache/pulsar#22155

Global LandingPage CSS class is introduced to keep the 18px font size on the "promotional" pages according to the original design. Maybe not the fully correct class name, but easy to understand and grep.